refactor : restructure backend utils helpers and auth
