+++ backend/db/models/user.js
@@ -1,8 +1,8 @@
 "use strict";
 const MetaModel = require("../MetaModel.js");
 const {Op} = require("sequelize");
-const {genSalt, genPwdHash, genPwd} = require("../../utils/auth.js");
-const {generateAnimalUsername} = require("../../utils/generator");
+const {genSalt, genPwdHash, genPwd} = require("../../webserver/auth/utils.js");
+const {generateAnimalUsername} = require("../../utils/helper/generator");
 const SequelizeSimpleCache = require("sequelize-simple-cache");
 
 module.exports = (sequelize, DataTypes) => {
