+++ backend/utils/helper/text.js
