+++ backend/webserver/sockets/study.js
@@ -1,5 +1,5 @@
 const Socket = require("../Socket.js");
-const {getEmailContent} = require("../../utils/emailHelper");
+const {getEmailContent} = require("../../utils/helper/email");
 
 /**
  * Handle all studies through websocket
